move methods for test data generation into separate class
